requirements.txt
configparser
asterisk-ami